Work on new canvas concept
Still very far from functional, but major code additions. The idea is to create a set of layers, each of which is invalidated separately. Then, all those layers get painted together as needed. This way, not everything needs to be repainted over and over. Instead, buffer bitmaps reside in each layer.